Relocate level_select_controller out of app/ui/ (#1285, refs #1193 OoS-C)
The file was misnamed — it contains seven free-function input event
handlers wired into the EnTT dispatcher, not an OOP controller. Move it
to app/systems/ (alongside its only consumer, input_dispatcher.cpp) and
rename to level_select_input_handlers to reflect what it actually is.
This honors the Fabian folder-layout invariant: only
components/ entities/ systems/ tags/ util/ live under app/ (the only
remaining offender is app/ui/ itself, tracked by #1193's larger
screen_controllers deletion).

Function signatures and behavior are unchanged. All 880 tests pass with
zero warnings.
