+++ docs/acceptance-traceability.md
@@ -25,7 +25,7 @@ The promotability gate that enforces this map lives in
 | `design-docs/feature-specs.md` Spec 2 Burnout Scoring | Burnout meter, zones, multiplier banking, burnout game-over, burnout chain rules | Superseded by issue #239 and the rhythm-first scoring docs (`design-docs/rhythm-design.md`, `design-docs/rhythm-spec.md`, `design-docs/energy-bar.md`). | Cut. Do not implement or gate release on these historical criteria. |
 | `design-docs/feature-specs.md` Spec 3 Obstacle Spawning & Difficulty | Distance-ahead timer spawning, piecewise speed ramp, cumulative obstacle unlocks, weighted-random lane assignment, combo frequency, burnout range shrink | Superseded by song-driven beatmaps and known content decisions in issues #328, #420, #441, and #446. Current beatmap/content gates live in `tools/validate_loop2_content_gates.py`, `tools/check_shape_change_gap.py`, and rhythm docs. | Deferred or superseded. Release gating uses shipped beatmap validators, not these historical random-spawner criteria. |
 | `design-docs/normalized-coordinates.md` | Normalized gameplay constants, collision/ring/test-player distances, beatmap positions, aspect-ratio enforcement | Display and virtual coordinate behavior is covered by `tests/test_ndc_viewport.cpp`, `tests/test_safe_area_layout.cpp`, and `tests/test_pr43_regression.cpp`; full gameplay-normalization migration remains parked by the document status. | Deferred. Not a TestFlight blocker until the feature is pulled from parked status. |
-| `design-docs/rguilayout-portable-c-integration.md` Validation Checklist | `.rgl`/generated header presence, generated-file discipline, controller wiring, raygui CMake integration, button dispatch, warning-free builds | Build and zero-warning policy are enforced by CMake/CI; controller behavior is covered by screen/controller tests such as `tests/test_components.cpp`, `tests/test_world_systems.cpp`, `tests/test_level_select_controller.cpp`, `tests/test_input_pipeline_behavior.cpp`, and generated layout compilation. | Active gate through build/tests; visual regression remains future scope as stated in that doc. |
+| `design-docs/rguilayout-portable-c-integration.md` Validation Checklist | `.rgl`/generated header presence, generated-file discipline, controller wiring, raygui CMake integration, button dispatch, warning-free builds | Build and zero-warning policy are enforced by CMake/CI; controller behavior is covered by screen/controller tests such as `tests/test_components.cpp`, `tests/test_world_systems.cpp`, `tests/test_level_select_input_handlers.cpp`, `tests/test_input_pipeline_behavior.cpp`, and generated layout compilation. | Active gate through build/tests; visual regression remains future scope as stated in that doc. |
 | `design-docs/beatmap-editor.md` section 6.1 task-file template | Placeholder checklist items `Item 1` through `Item 3` | These are template placeholders for sub-agent task files, not project acceptance criteria. | Not applicable. Do not count as release criteria. |
 | `docs/testflight-readiness.md` unchecked items | MetricKit implementation, privacy/ATT verification, audio pipeline follow-up, tester group/public link | Each line already names its owning issue or operational owner in the readiness plan. | Active TestFlight checklist items; gate through section 3.5 and the linked issues. |
